+++ CHANGELOG.md
@@ -11,3 +11,4 @@
 - Rename `BBands.history` to `BBands.graph`
 - Rename `IndicatorGraphHandler.process_graph_for_existing` to `process_graph`
 - Rename `IndicatorSignalHandler.process_signals_for_existing` to `process_signals`
+- Create errors for signal / graph processing